--- First Plugin/src/main/resources/plugin.yml
@@ -1,14 +1,0 @@
-name: FirstPlugin
-version: '${project.version}'
-main: coolperson.firstplugin.Main
-api-version: 1.18
-website: coolperson.me
-
-commands:
-  rules:
-    description: shows the player the rules
-    usage: /rules
-
-  warp:
-    description: Warps the player to a location.
-    usage: /warp
